Add prompt fraction to tree

+++ LArIATRecoAlg/OpHitBuilderAlg.h
@@ -42,48 +42,48 @@ class OpHitBuilderAlg{
 
   std::vector<short>              GetHits( std::vector<short>);
   std::vector<short>              HitMerger( std::vector<short>, short, int);
-  std::vector<double>             MakeGradient( std::vector<short> );
-  std::vector<double>             GetBaselineAndRMS( std::vector<short>, short, short);
-  std::vector<double>               GetBaselineAndRMS( std::vector<double>, short, short);
-  std::vector<double>             GetHitInfo( std::vector<short>, short);
-  double                          GetHitAmplitude( std::vector<short>, short);
-  double                          GetHitPromptIntegral( std::vector<short>, short);
-  double                          GetHitFullIntegral(   std::vector<short>, short);  
+  std::vector<float>             MakeGradient( std::vector<short> );
+  std::vector<float>             GetBaselineAndRMS( std::vector<short>, short, short);
+  std::vector<float>               GetBaselineAndRMS( std::vector<float>, short, short);
+  std::vector<float>             GetHitInfo( std::vector<short>, short);
+  float                          GetHitAmplitude( std::vector<short>, short);
+  float                          GetHitPromptIntegral( std::vector<short>, short);
+  float                          GetHitFullIntegral(   std::vector<short>, short);  
   bool                              IsCleanBeamWaveform( raw::OpDetPulse );
   short                             GetLocalMinimum( std::vector<short>, short);
-  double                          GetLocalMinimum( std::vector<double>, short);
+  float                          GetLocalMinimum( std::vector<float>, short);
   
   // Average waveform vector
-  std::vector<double>   AverageWaveform;
+  std::vector<float>   AverageWaveform;
   int                   AverageWaveform_count;
   int                   AddHitToAverageWaveform;
 
   // Fit parameters
-  double prepulse_baseline;
-  double prepulse_rms;
-  double fit_FastNorm;
-  double fit_FastTau;
-  double fit_SlowNorm;
-  double fit_SlowTau;
-  double fit_ReducedChi2;
+  float prepulse_baseline;
+  float prepulse_rms;
+  float fit_FastNorm;
+  float fit_FastTau;
+  float fit_SlowNorm;
+  float fit_SlowTau;
+  float fit_ReducedChi2;
 
   bool bVerbose;
   bool      fUsePrepulseFit;
 
  private:
   
-  double  fGradientHitThreshold;
-  double  fPulseHitThreshold;
-  double  fGradientRMSFilterThreshold;
-  double  fMinHitSeparation;  
+  float  fGradientHitThreshold;
+  float  fPulseHitThreshold;
+  float  fGradientRMSFilterThreshold;
+  float  fMinHitSeparation;  
   short     fBaselineWindowLength;
   short     fPrePulseBaselineFit;
   short     fPromptWindowLength;
   short     fFullWindowLength;
-  double    fMvPerADC;
-  double    fTimestampCut;
-  double    fPrePulseTau1;
-  double    fPrePulseTau2;
+  float    fMvPerADC;
+  float    fTimestampCut;
+  float    fPrePulseTau1;
+  float    fPrePulseTau2;
   
 };
